requirements.txt
Django==1.8
django-braces==1.8.1
django-allauth==0.25.2
django-crispy-forms==1.6.0
Pillow==3.2.0
pinax-notifications==3.0.1
